projectman/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
chardet==5.2.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==5.0.1
django-ckeditor==6.7.0
django-js-asset==2.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
mysqlclient==2.2.1
outcome==1.3.0.post0
pillow==10.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
PySocks==1.7.1
redis==5.0.1
reportlab==4.0.9
selenium==4.16.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.1.0
wsproto==1.2.0
zope.interface==6.1